Use rectByAccountingForStyleSize to get the true dimensions of objects with outside borders, shadows, blurs and reflections.
This exports the images at the right size, but there's still a problem when positioning them on the canvas: we need to offset the image position to take effects into account.

Right now the offset calculation assumes that effects are symmetrical in both axes, and will fail miserably for effects that aren't (your typical drop shadow, or a basic reflection, for example).

Anyway, I will go out on a limb and declare: this commit fixes #7, and I will open a new issue (#13, most probably) for the offset calculation.

Infinite thanks to @pieteromvlee for his help cracking this one.